requirements.txt
anyio==4.0.0
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==4.2.4
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.0
construct==2.10.68
construct-typing==0.5.6
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
greenlet==3.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
install==1.3.5
ipykernel==6.26.0
ipython==8.17.2
isodate==0.6.1
jedi==0.19.1
jsonalias==0.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
outcome==1.3.0.post0
owlready2==0.45
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
rdflib==7.0.0
rel==0.4.9
requests==2.31.0
rfc3986==1.5.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
solders==0.18.1
sortedcontainers==2.4.0
soupsieve==2.5
sparql-dataframe==0.4
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.23
stack-data==0.6.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trio==0.23.1
trio-websocket==0.11.1
types-cachetools==4.2.10
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.8
urllib3==2.1.0
w3lib==2.1.2
wcwidth==0.2.10
websocket-client==1.6.4
websockets==11.0.3
wsproto==1.2.0